Fix and clarify test model loading
