+++ spec/internal/app/models/altered_inheriting_taggable_model.rb
@@ -1,3 +1,5 @@
+require_relative 'taggable_model'
+
 class AlteredInheritingTaggableModel < TaggableModel
   acts_as_taggable_on :parts
 end
