+++ spec/internal/app/models/company.rb
