+++ spec/internal/app/models/inheriting_taggable_model.rb
@@ -1,2 +1,4 @@
+require_relative 'taggable_model'
+
 class InheritingTaggableModel < TaggableModel
 end
