+++ spec/internal/app/models/market.rb
