+++ spec/internal/app/models/non_standard_id_taggable_model.rb
@@ -1,5 +1,5 @@
 class NonStandardIdTaggableModel < ActiveRecord::Base
-  self.primary_key = 'an_id'
+  self.primary_key = :an_id
   acts_as_taggable
   acts_as_taggable_on :languages
   acts_as_taggable_on :skills
