+++ spec/internal/app/models/student.rb
@@ -1,2 +1,4 @@
+require_relative 'user'
+
 class Student < User
 end
