+++ spec/internal/app/models/user.rb
