+++ spec/support/database.rb
@@ -28,8 +28,8 @@ if File.exist?(database_yml)
     ActiveRecord::Base.establish_connection(config)
   end
 
-  load(File.dirname(__FILE__) + '/../internal/db/schema.rb')
-  load(File.dirname(__FILE__) + '/../internal/app/models/models.rb')
+  require File.dirname(__FILE__) + '/../internal/db/schema.rb'
+  Dir[File.dirname(__dir__) + '/internal/app/models/*.rb'].each { |f| require f }
 
 else
   fail "Please create #{database_yml} first to configure your database. Take a look at: #{database_yml}.sample"
